+++ chalice/cli/__init__.py
@@ -5,12 +5,14 @@ Contains commands for deploying chalice.
 """
 import os
 import json
+import logging
 
 import click
 import botocore.exceptions
 import botocore.session
 
 from chalice import deployer
+from chalice import __version__ as chalice_version
 from chalice.logs import LogRetriever
 from chalice import prompts
 from chalice.config import Config
@@ -51,11 +53,19 @@ def index():
 """
 
 
+def create_botocore_session(profile=None, debug=False):
+    session = botocore.session.Session(profile=profile)
+    session.user_agent_extra = 'chalice/%s' % chalice_version
+    if debug:
+        session.set_debug_logger('')
+        inject_large_request_body_filter()
+    return session
+
+
 def show_lambda_logs(config, max_entries, include_lambda_messages):
-    import botocore.session
     lambda_arn = config.lambda_arn
     profile = config.profile
-    client = botocore.session.Session(profile=profile).create_client('logs')
+    client = create_botocore_session(profile).create_client('logs')
     retriever = LogRetriever.create_from_arn(client, lambda_arn)
     events = retriever.retrieve_logs(
         include_lambda_messages=include_lambda_messages,
@@ -91,7 +101,27 @@ def load_chalice_app(project_dir):
         return g['app']
 
 
+def inject_large_request_body_filter():
+    log = logging.getLogger('botocore.endpoint')
+    log.addFilter(LargeRequestBodyFilter())
+
+
+class LargeRequestBodyFilter(logging.Filter):
+    def filter(self, record):
+        if record.msg.startswith('Making request'):
+            if record.args[0].name in ['UpdateFunctionCode', 'CreateFunction']:
+                # When using the ZipFile argument (which is used in chalice),
+                # the entire deployment package zip is sent as a base64 encoded
+                # string.  We don't want this to clutter the debug logs
+                # so we don't log the request body for lambda operations
+                # that have the ZipFile arg.
+                record.args = (record.args[:-1] +
+                               ('(... omitted from logs due to size ...)',))
+        return True
+
+
 @click.group()
+@click.version_option(version=chalice_version, message='%(prog)s %(version)s')
 @click.pass_context
 def cli(ctx):
     pass
@@ -110,9 +140,12 @@ def local(ctx):
               default=True,
               help='Automatically generate IAM policy for app code.')
 @click.option('--profile', help='Override profile at deploy time.')
+@click.option('--debug/--no-debug',
+              default=False,
+              help='Print debug logs to stderr.')
 @click.argument('stage', nargs=1, required=False)
 @click.pass_context
-def deploy(ctx, project_dir, autogen_policy, profile, stage):
+def deploy(ctx, project_dir, autogen_policy, profile, debug, stage):
     user_provided_params = {}
     default_params = {}
     if project_dir is None:
@@ -136,7 +169,7 @@ def deploy(ctx, project_dir, autogen_policy, profile, stage):
     if profile:
         user_provided_params['profile'] = profile
     config = Config(user_provided_params, config_from_disk, default_params)
-    session = botocore.session.Session(profile=config.profile)
+    session = create_botocore_session(profile=config.profile, debug=debug)
     d = deployer.create_default_deployer(session=session, prompter=click)
     try:
         d.deploy(config)
@@ -146,11 +179,6 @@ def deploy(ctx, project_dir, autogen_policy, profile, stage):
                                  "environment variable or set the "
                                  "region value in our ~/.aws/config file.")
         e.exit_code = 2
-        raise e
-    except Exception as e:
-        raise
-        e = click.ClickException("Error when deploying: %s" % e)
-        e.exit_code = 1
         raise e